[fix] 스나이퍼 라이플 Dummy 추가(만약 Sniper Rifle 시 Ammo만 증가)

+++ Assets/6.Debug/CustomDebugWindow.cs
@@ -114,6 +114,22 @@
                 AssetDatabase.SaveAssets();
             }
 
+            if (GUILayout.Button("Get All Available Parts"))
+            {
+                playerObj = GameObject.FindWithTag("Player");
+                if (!playerObj.TryGetComponent(out Player player)) return;
+                
+                player.PlayerWeapon.GetAllAvailableParts();
+            }
+
+            if (GUILayout.Button("Forge Weapon"))
+            {
+                playerObj = GameObject.FindWithTag("Player");
+                if (!playerObj.TryGetComponent(out Player player)) return;
+
+                player.PlayerWeapon.ForgeWeapon();
+            }
+            
             if (GUILayout.Button("Clear PlayerPrefs"))
             {
                 PlayerPrefs.DeleteAll();
